launchflow-endpoint-template/launchflow_endpoint_template/settings.py
"""This file defines any environment variables needed to run."""

from enum import Enum
import os

import dotenv


class Environment(Enum):
    DEV = "dev"
    PROD = "prod"
    LOCAL = "local"


class Settings:
    def __init__(self) -> None:
        dotenv.load_dotenv()
        self.env = Environment(os.getenv("STACK", "local"))


env = Settings()
