launchflow-endpoint-template/requirements.txt
buildflow @ git+https://github.com/launchflow/buildflow.git
python-dotenv
ray==2.7.0
